refactor: split dataset list into subcomponents
