+++ packages_rs/nextclade-web/src/state/autodetect.state.ts
@@ -1,9 +1,8 @@
 /* eslint-disable no-loops/no-loops */
-import unique from 'fork-ts-checker-webpack-plugin/lib/utils/array/unique'
-import { isEmpty, isNil } from 'lodash'
+import { isEmpty, isNil, uniq } from 'lodash'
 import { atom, atomFamily, DefaultValue, selector, selectorFamily } from 'recoil'
-import type { MinimizerIndexJson, MinimizerSearchRecord } from 'src/types'
 import { isDefaultValue } from 'src/state/utils/isDefaultValue'
+import type { MinimizerIndexJson, MinimizerSearchRecord } from 'src/types'
 
 export const minimizerIndexAtom = atom<MinimizerIndexJson>({
   key: 'minimizerIndexAtom',
@@ -51,7 +50,7 @@ export const autodetectResultByIndexAtom = selectorFamily<MinimizerSearchRecord,
 export const DATASET_ID_UNDETECTED = 'undetected'
 
 export function groupByDatasets(records: MinimizerSearchRecord[]): Record<string, MinimizerSearchRecord[]> {
-  const names = unique(records.flatMap((record) => record.result.datasets.map((dataset) => dataset.name)))
+  const names = uniq(records.flatMap((record) => record.result.datasets.map((dataset) => dataset.name)))
   let byDataset = {}
   for (const name of names) {
     const selectedRecords = records.filter((record) => record.result.datasets.some((dataset) => dataset.name === name))
